Fix branching
Il faut vérifier si il reste des variables à assigner avant de faire
des calculs de densité pour une contrainte.

+++ problems/Sudoku.cpp
@@ -176,7 +176,7 @@ public:
         } else if (opt.branching() == BRANCH_AFC) {
             branch(*this, x, INT_VAR_AFC_MAX(opt.decay()), INT_VAL_SPLIT_MIN());
         } else if (opt.branching() == BRANCH_CBS) {
-            cbsbranch(*this, constraints, std::greater<int>());
+            cbsbranch(*this, constraints, std::greater<double>());
         }
     }
